refactor!: move various types to a new namespace named 'Carbonate.Core'
