+++ Carbonate/Core/IReceiveReactor.cs
@@ -2,7 +2,7 @@
 // Copyright (c) KinsonDigital. All rights reserved.
 // </copyright>
 
-namespace Carbonate;
+namespace Carbonate.Core;
 
 /// <summary>
 /// A reactor capable of standard functionality that can receive push notifications.
