+++ Carbonate/Core/IResult.cs
@@ -1,13 +1,13 @@
-﻿// <copyright file="IMessage.cs" company="KinsonDigital">
+﻿// <copyright file="IResult.cs" company="KinsonDigital">
 // Copyright (c) KinsonDigital. All rights reserved.
 // </copyright>
 
-namespace Carbonate;
+namespace Carbonate.Core;
 
 /// <summary>
-/// A message that can be passed containing the data to consume.
+/// Contains data that is returned from a <see cref="IRespondReactor"/> and <see cref="PullReactable"/>.
 /// </summary>
-public interface IMessage
+public interface IResult
 {
     /// <summary>
     /// Gets the data as the type <typeparamref name="T"/>.
@@ -15,6 +15,6 @@ public interface IMessage
     /// <param name="onError">The action to invoke if an exception occurs.</param>
     /// <typeparam name="T">The type to deserialize the message into.</typeparam>
     /// <returns>The deserialized message data.</returns>
-    public T? GetData<T>(Action<Exception>? onError = null)
+    public T? GetValue<T>(Action<Exception>? onError = null)
         where T : class;
 }
