+++ Carbonate/ResultFactory.cs
@@ -5,6 +5,7 @@
 namespace Carbonate;
 
 using System.Diagnostics.CodeAnalysis;
+using Core;
 using Services;
 
 /// <summary>
