+++ Testing/CarbonateTests/PushReactableTests.cs
@@ -7,6 +7,7 @@ namespace CarbonateTests;
 
 using System.Text.Json;
 using Carbonate;
+using Carbonate.Core;
 using Carbonate.Services;
 using Helpers;
 using FluentAssertions;
